requirements.txt
# Python 3.11.9

opencv-python==4.12.0
numpy==2.2.6
pycozmo==0.8.0
Pillow==12.0.0

# Deep learning
torch==2.9.1+cpu
transformers==4.57.3
ultralytics==8.3.241
